requirements.txt
certifi==2019.9.11
chardet==3.0.4
fhir.resources==5.0.1
idna==2.8
isodate==0.6.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
